# packages/shared-types/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RuleCreate": {
        "additionalProperties": false,
        "properties": {
          "condition": {
            "additionalProperties": true,
            "description": "Condition AST for the first draft.",
            "title": "Condition",
            "type": "object"
          },
          "key": {
            "maxLength": 100,
            "pattern": "^[a-z][a-z0-9_]*(\\.[a-z][a-z0-9_]*)*$",
            "title": "Key",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "key",
          "title",
          "condition"
        ],
        "title": "RuleCreate",
        "type": "object"
      },
...
      "RuleSetCreate": {
        "additionalProperties": false,
        "properties": {
          "applies_when": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Condition AST; omit when the rule set applies everywhere.",
            "title": "Applies When"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "jurisdiction": {
            "pattern": "^(CTH|QLD|NSW|VIC|SA|WA|TAS|NT|ACT|LGA:[A-Z][A-Z0-9_]{1,59})$",
            "title": "Jurisdiction",
            "type": "string"
          },
          "key": {
            "maxLength": 100,
            "pattern": "^[a-z][a-z0-9_]*(\\.[a-z][a-z0-9_]*)*$",
            "title": "Key",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "vertical": {
            "$ref": "#/components/schemas/Vertical"
          }
        },
        "required": [
          "key",
          "vertical",
          "jurisdiction",
          "title"
        ],
        "title": "RuleSetCreate",
        "type": "object"
      },
      "RuleSetOut": {
        "properties": {
          "applies_when": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applies When"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "jurisdiction": {
            "title": "Jurisdiction",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/RuleSummary"
            },
            "title": "Rules",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "vertical": {
            "$ref": "#/components/schemas/Vertical"
          }
        },
        "required": [
          "id",
          "key",
          "vertical",
          "jurisdiction",
          "title",
          "description",
          "applies_when",
          "rules",
          "created_at",
          "updated_at"
        ],
        "title": "RuleSetOut",
        "type": "object"
      },
...
      "RuleSetUpdate": {
        "additionalProperties": false,
        "properties": {
          "applies_when": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applies When"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "jurisdiction": {
            "anyOf": [
              {
                "pattern": "^(CTH|QLD|NSW|VIC|SA|WA|TAS|NT|ACT|LGA:[A-Z][A-Z0-9_]{1,59})$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jurisdiction"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "RuleSetUpdate",
        "type": "object"
      },
...
      "Vertical": {
        "enum": [
          "PLANNING",
          "VESSEL",
          "BUSINESS",
          "GRANT",
          "SELL",
          "RENT"
        ],
        "title": "Vertical",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "ApprovalReady API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/rule-sets": {
      "get": {
        "operationId": "list_rule_sets_v1_admin_rule_sets_get",
        "parameters": [
          {
            "in": "query",
            "name": "vertical",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Vertical"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vertical"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RuleSetOut"
                  },
                  "title": "Response List Rule Sets V1 Admin Rule Sets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Rule Sets",
        "tags": [
          "admin: rules"
        ]
      },
      "post": {
        "operationId": "create_rule_set_v1_admin_rule_sets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleSetCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Rule Set",
        "tags": [
          "admin: rules"
        ]
      }
    },
    "/v1/admin/rule-sets/{rule_set_id}": {
      "get": {
        "operationId": "get_rule_set_v1_admin_rule_sets__rule_set_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "rule_set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Set Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Rule Set",
        "tags": [
          "admin: rules"
        ]
      },
      "patch": {
        "operationId": "update_rule_set_v1_admin_rule_sets__rule_set_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "rule_set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Set Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleSetUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleSetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Rule Set",
        "tags": [
          "admin: rules"
        ]
      }
    },
    "/v1/admin/rule-sets/{rule_set_id}/rules": {
      "post": {
        "description": "Create a rule with its first draft version.",
        "operationId": "create_rule_v1_admin_rule_sets__rule_set_id__rules_post",
        "parameters": [
          {
            "in": "path",
            "name": "rule_set_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Set Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Rule",
        "tags": [
          "admin: rules"
        ]
      }
    },
    "/v1/admin/rule-versions/{version_id}": {
      "delete": {
        "operationId": "delete_rule_version_v1_admin_rule_versions__version_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Rule Version",
        "tags": [
          "admin: rules"
        ]
      },
      "get": {
        "operationId": "get_rule_version_v1_admin_rule_versions__version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleVersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Rule Version",
        "tags": [
          "admin: rules"
        ]
      },
      "put": {
        "operationId": "save_rule_version_v1_admin_rule_versions__version_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleVersionContent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleVersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Rule Version",
        "tags": [
          "admin: rules"
        ]
      }
    },
    "/v1/admin/rule-versions/{version_id}/checks": {
      "get": {
        "description": "Run the publish gate (including every test case) without publishing.",
        "operationId": "check_rule_version_v1_admin_rule_versions__version_id__checks_get",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishChecksOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check Rule Version",
        "tags": [
          "admin: rules"
        ]
      }
    },
    "/v1/admin/rule-versions/{version_id}/evaluate": {
      "post": {
        "description": "Try the version against some facts, with the confidence its sources allow today.",
        "operationId": "evaluate_rule_version_v1_admin_rule_versions__version_id__evaluate_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate Rule Version",
        "tags": [
          "admin: rules"
        ]
      }
    },
    "/v1/admin/rule-versions/{version_id}/publish": {
      "post": {
        "operationId": "publish_rule_version_v1_admin_rule_versions__version_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleVersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish Rule Version",
        "tags": [
          "admin: rules"
        ]
      }
    },
    "/v1/admin/rule-versions/{version_id}/retire": {
      "post": {
        "operationId": "retire_rule_version_v1_admin_rule_versions__version_id__retire_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleVersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retire Rule Version",
        "tags": [
          "admin: rules"
        ]
      }
    },
    "/v1/admin/rules/{rule_id}": {
      "get": {
        "operationId": "get_rule_v1_admin_rules__rule_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Rule",
        "tags": [
          "admin: rules"
        ]
      },
      "patch": {
        "operationId": "update_rule_v1_admin_rules__rule_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Rule",
        "tags": [
          "admin: rules"
        ]
      }
    },
    "/v1/admin/rules/{rule_id}/versions": {
      "post": {
        "description": "Start a new draft, copied from the latest version.",
        "operationId": "create_rule_version_v1_admin_rules__rule_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleVersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Rule Version",
        "tags": [
          "admin: rules"
        ]
      }
    },
...
  }
}